CIE plot

Plot a CIE (XYZ, LAB, or LCH) chromaticity diagram.

cieplot(ciedata, mono = TRUE, out.lwd = NULL, out.lcol = "black",
  out.lty = 1, theta = 45, phi = 10, r = 1e+06, zoom = 1,
  box = FALSE, margin = c(0, 0, 0, 0), view, scale.y, axis, grid, ...)

Arguments

ciedata
(required). a data frame, possibly a result from the colspace or cie function, containing values for 'x', 'y' and 'z' coordinates for the CIEXYZ model, or LAB coordinates for the CIELAB (or CIELCh models), as columns (labeled as such).

mono
should the monochromatic loci (the 'horseshoe') be plotted when space = 'ciexyz'? Defaults to TRUE.

out.lwd, out.lcol, out.lty
graphical parameters for the monochromatic loci outline.

theta
angle to rotate the plot in the xy plane when space = 'cielab' (defaults to 10).

phi
angle to rotate the plot in the yz plane when space = 'cielab' (defaults to 45).

r
the distance of the eyepoint from the centre of the plotting box when space = 'cielab'. Very high values approximate an orthographic projection (defaults to 1e6). See persp for details. . See persp for details.

zoom
zooms in (values greater than 1) or out (values between 0 and 1) from the plotting area when space = 'cielab'.

box
logical. Should the plot area box and axes be plotted? (defaults to FALSE)

margin
vector of four numbers specifying drawing margins for CIELAB plot (defaults to c(0,0,0,0))

view,
scale.y, axis, grid deprecated arguments.

...
Additional graphical options. See par.

References

Smith T, Guild J. (1932) The CIE colorimetric standards and their use. Transactions of the Optical Society, 33(3), 73-134.

Westland S, Ripamonti C, Cheung V. (2012). Computational colour science using MATLAB. John Wiley & Sons.

Examples

# NOT RUN {
data(flowers)

# CIEXYZ
vis.flowers <- vismodel(flowers, visual = 'cie10', illum = 'D65', vonkries = TRUE, relative = FALSE)
xyz.flowers <- colspace(vis.flowers, space = 'ciexyz')
plot(xyz.flowers)

#CIELAB
lab.flowers <- colspace(vis.flowers, space = 'cielab')
plot(lab.flowers)
# }