Import spectra files

Finds and imports spectra files from a folder. Currently works for reflectance files generated in Ocean Optics SpectraSuite (USB2000, USB4000 and Jaz spectrometers), CRAIC software (after exporting) and Avantes (before or after exporting).

getspec(where = getwd(), ext = "txt", lim = c(300, 700),
  decimal = ".", sep = NULL, subdir = FALSE, subdir.names = FALSE,
  cores = getOption("mc.cores", 2L), ignore.case = TRUE)

Arguments

where
(required) folder in which files are located.

ext
file extension to be searched for, without the "." (defaults to "txt").

lim
a vector with two numbers determining the wavelength limits to be considered (defaults to 300 and 700).

decimal
character to be used to identify decimal plates (defaults to ".").

sep
column delimiting characters to be considered in addition to the default (which are: tab, space, and ";")

subdir
should subdirectories within the where folder be included in the search? (defaults to FALSE).

subdir.names
should subdirectory path be included in the name of the spectra? (defaults to FALSE).

cores
Number of cores to be used. If greater than 1, import will use parallel processing (not available in Windows).

ignore.case
Logical. Should the extension search be case insensitive? (defaults to TRUE)

Value

A data frame, of class rspec, containing individual imported spectral files as columns. Reflectance values are interpolated to the nearest wavelength integer.

References

Montgomerie R (2006) Analyzing colors. In: Hill G, McGraw K (eds) Bird coloration. Harvard University Press, Cambridge, pp 90-147.

Examples

getspec(system.file("testdata", package = "pavo"), lim = c(400, 900))

#> 2 files found; importing spectra:

#> Warning: Could not import one or more files:
#> /Users/twhite/Dropbox (Sydney Uni)/notes/research/projects/pavo/inst/testdata/failfile.txt

#> 
#> The spectral data contain 27 negative value(s), which may produce unexpected results if used in models. Consider using procspec() to correct them.

#>      wl   OceanView
#> 1   400 -1.85001333
#> 2   401 -1.83632667
#> 3   402 -1.81976667
#> 4   403 -1.87096000
#> 5   404 -1.85432889
#> 6   405 -1.79413182
#> 7   406 -1.71129778
#> 8   407 -1.66359333
#> 9   408 -1.59773636
#> 10  409 -1.51895556
#> 11  410 -1.38170000
#> 12  411 -1.34129556
#> 13  412 -1.21987333
#> 14  413 -1.13581818
#> 15  414 -0.97279200
#> 16  415 -0.87202864
#> 17  416 -0.72690400
#> 18  417 -0.60824909
#> 19  418 -0.50232667
#> 20  419 -0.41930778
#> 21  420 -0.34881977
#> 22  421 -0.26366622
#> 23  422 -0.18924533
#> 24  423 -0.18617773
#> 25  424 -0.12562622
#> 26  425 -0.09831220
#> 27  426 -0.03143027
#> 28  427  0.02550981
#> 29  428  0.05055975
#> 30  429  0.13381311
#> 31  430  0.21788533
#> 32  431  0.31900911
#> 33  432  0.45502111
#> 34  433  0.59742533
#> 35  434  0.69955333
#> 36  435  0.75039500
#> 37  436  0.81478400
#> 38  437  0.94077111
#> 39  438  1.03958889
#> 40  439  1.09107333
#> 41  440  1.15579091
#> 42  441  1.21556364
#> 43  442  1.22850909
#> 44  443  1.23306136
#> 45  444  1.28651778
#> 46  445  1.30096818
#> 47  446  1.31453636
#> 48  447  1.34783182
#> 49  448  1.33573636
#> 50  449  1.33644545
#> 51  450  1.34316889
#> 52  451  1.33907333
#> 53  452  1.37330000
#> 54  453  1.45070909
#> 55  454  1.54020000
#> 56  455  1.63002000
#> 57  456  1.66520000
#> 58  457  1.71887500
#> 59  458  1.80837273
#> 60  459  1.87730000
#> 61  460  1.95645000
#> 62  461  2.04210000
#> 63  462  2.10343182
#> 64  463  2.11889318
#> 65  464  2.17003864
#> 66  465  2.19886364
#> 67  466  2.26284545
#> 68  467  2.29114773
#> 69  468  2.33358409
#> 70  469  2.35570000
#> 71  470  2.33690909
#> 72  471  2.31667727
#> 73  472  2.29979091
#> 74  473  2.29602727
#> 75  474  2.29489545
#> 76  475  2.26976000
#> 77  476  2.23601818
#> 78  477  2.25004545
#> 79  478  2.24700000
#> 80  479  2.27552727
#> 81  480  2.31204545
#> 82  481  2.36020909
#> 83  482  2.40431818
#> 84  483  2.43608182
#> 85  484  2.49750909
#> 86  485  2.53264545
#> 87  486  2.57912727
#> 88  487  2.63494545
#> 89  488  2.68787273
#> 90  489  2.74460000
#> 91  490  2.77756364
#> 92  491  2.82850000
#> 93  492  2.88741818
#> 94  493  2.91897955
#> 95  494  3.00185227
#> 96  495  3.08054773
#> 97  496  3.14864773
#> 98  497  3.18639091
#> 99  498  3.23610000
#> 100 499  3.30177727
#> 101 500  3.34482500
#> 102 501  3.40994091
#> 103 502  3.48407273
#> 104 503  3.53668182
#> 105 504  3.55827727
#> 106 505  3.58503864
#> 107 506  3.63969091
#> 108 507  3.68935116
#> 109 508  3.73627500
#> 110 509  3.78623256
#> 111 510  3.81267273
#> 112 511  3.85621818
#> 113 512  3.87510000
#> 114 513  3.92130000
#> 115 514  3.93744545
#> 116 515  3.98348182
#> 117 516  4.00478140
#> 118 517  4.04522093
#> 119 518  4.07917674
#> 120 519  4.11931395
#> 121 520  4.13809091
#> 122 521  4.17230698
#> 123 522  4.22787727
#> 124 523  4.30892045
#> 125 524  4.37150000
#> 126 525  4.43750930
#> 127 526  4.51088605
#> 128 527  4.56162558
#> 129 528  4.62346136
#> 130 529  4.67360000
#> 131 530  4.72718636
#> 132 531  4.76613256
#> 133 532  4.80919767
#> 134 533  4.81907209
#> 135 534  4.84860000
#> 136 535  4.87852093
#> 137 536  4.91514884
#> 138 537  4.94343636
#> 139 538  4.99976977
#> 140 539  5.05271818
#> 141 540  5.10200000
#> 142 541  5.11103636
#> 143 542  5.14665349
#> 144 543  5.18173488
#> 145 544  5.20706977
#> 146 545  5.22666977
#> 147 546  5.25981395
#> 148 547  5.26513864
#> 149 548  5.29246512
#> 150 549  5.31376047
#> 151 550  5.36022093
#> 152 551  5.41352558
#> 153 552  5.47350000
#> 154 553  5.52821628
#> 155 554  5.58417273
#> 156 555  5.64797674
#> 157 556  5.72082326
#> 158 557  5.79423256
#> 159 558  5.87380930
#> 160 559  5.96296047
#> 161 560  6.03876977
#> 162 561  6.10856047
#> 163 562  6.18718140
#> 164 563  6.28380909
#> 165 564  6.35669302
#> 166 565  6.44931395
#> 167 566  6.53550000
#> 168 567  6.64226744
#> 169 568  6.73753023
#> 170 569  6.83254186
#> 171 570  6.94456744
#> 172 571  7.05558140
#> 173 572  7.16228372
#> 174 573  7.28368140
#> 175 574  7.39260930
#> 176 575  7.48976512
#> 177 576  7.58182093
#> 178 577  7.66536047
#> 179 578  7.77443023
#> 180 579  7.86103256
#> 181 580  7.97275581
#> 182 581  8.07847209
#> 183 582  8.17898095
#> 184 583  8.26615814
#> 185 584  8.37995814
#> 186 585  8.50337907
#> 187 586  8.61043256
#> 188 587  8.73240465
#> 189 588  8.87738837
#> 190 589  8.95826977
#> 191 590  9.06186744
#> 192 591  9.14616905
#> 193 592  9.25029535
#> 194 593  9.34788571
#> 195 594  9.40765116
#> 196 595  9.50761163
#> 197 596  9.57407381
#> 198 597  9.62028333
#> 199 598  9.70728140
#> 200 599  9.76580000
#> 201 600  9.83593333
#> 202 601  9.89744048
#> 203 602  9.95184048
#> 204 603 10.02144186
#> 205 604 10.05025581
#> 206 605 10.08241860
#> 207 606 10.11400000
#> 208 607 10.14325581
#> 209 608 10.19609302
#> 210 609 10.23469767
#> 211 610 10.28072093
#> 212 611 10.31066667
#> 213 612 10.35528571
#> 214 613 10.42714286
#> 215 614 10.49383721
#> 216 615 10.58767442
#> 217 616 10.66360465
#> 218 617 10.72957143
#> 219 618 10.80558140
#> 220 619 10.88100000
#> 221 620 10.95921429
#> 222 621 11.01948837
#> 223 622 11.05971429
#> 224 623 11.09655814
#> 225 624 11.09519048
#> 226 625 11.12451163
#> 227 626 11.13319048
#> 228 627 11.13452381
#> 229 628 11.13848837
#> 230 629 11.11695238
#> 231 630 11.09409524
#> 232 631 11.06900000
#> 233 632 11.04471429
#> 234 633 11.04471429
#> 235 634 11.01032558
#> 236 635 10.99657143
#> 237 636 10.98833333
#> 238 637 10.96590476
#> 239 638 10.96781395
#> 240 639 10.97114286
#> 241 640 10.98633333
#> 242 641 11.00097619
#> 243 642 11.00392857
#> 244 643 11.02560465
#> 245 644 11.04185714
#> 246 645 11.08652381
#> 247 646 11.11285714
#> 248 647 11.14341860
#> 249 648 11.18985714
#> 250 649 11.24028571
#> 251 650 11.28050000
#> 252 651 11.32521429
#> 253 652 11.37633333
#> 254 653 11.40502381
#> 255 654 11.42657143
#> 256 655 11.47295238
#> 257 656 11.52519048
#> 258 657 11.57078571
#> 259 658 11.61185714
#> 260 659 11.67133333
#> 261 660 11.71439024
#> 262 661 11.78500000
#> 263 662 11.85300000
#> 264 663 11.92509524
#> 265 664 12.00495238
#> 266 665 12.07283333
#> 267 666 12.13685714
#> 268 667 12.19492857
#> 269 668 12.25700000
#> 270 669 12.32085714
#> 271 670 12.36685714
#> 272 671 12.42850000
#> 273 672 12.47085714
#> 274 673 12.50951220
#> 275 674 12.54259524
#> 276 675 12.59585714
#> 277 676 12.63980952
#> 278 677 12.65759524
#> 279 678 12.69900000
#> 280 679 12.72400000
#> 281 680 12.73914286
#> 282 681 12.74304762
#> 283 682 12.76741463
#> 284 683 12.79014286
#> 285 684 12.78368293
#> 286 685 12.79258537
#> 287 686 12.81997561
#> 288 687 12.82111905
#> 289 688 12.82258537
#> 290 689 12.84931707
#> 291 690 12.86458537
#> 292 691 12.87402439
#> 293 692 12.88634146
#> 294 693 12.90200000
#> 295 694 12.93116667
#> 296 695 12.93243902
#> 297 696 12.92887805
#> 298 697 12.93107317
#> 299 698 12.90170732
#> 300 699 12.87670732
#> 301 700 12.85007143
#> 302 701 12.83146341
#> 303 702 12.79523810
#> 304 703 12.73882927
#> 305 704 12.69443902
#> 306 705 12.66456098
#> 307 706 12.63575610
#> 308 707 12.61853659
#> 309 708 12.57885714
#> 310 709 12.55819512
#> 311 710 12.51519512
#> 312 711 12.50776190
#> 313 712 12.47275610
#> 314 713 12.44480488
#> 315 714 12.41765854
#> 316 715 12.36680488
#> 317 716 12.33664286
#> 318 717 12.29195122
#> 319 718 12.25292857
#> 320 719 12.22648780
#> 321 720 12.19907143
#> 322 721 12.16956098
#> 323 722 12.14424390
#> 324 723 12.11400000
#> 325 724 12.07841463
#> 326 725 12.04875610
#> 327 726 12.04324390
#> 328 727 12.04736585
#> 329 728 12.01395122
#> 330 729 12.00197561
#> 331 730 11.99212195
#> 332 731 12.00456098
#> 333 732 12.01285366
#> 334 733 12.03217073
#> 335 734 12.05017073
#> 336 735 12.07497561
#> 337 736 12.07741463
#> 338 737 12.08690244
#> 339 738 12.06343902
#> 340 739 12.08263415
#> 341 740 12.09056098
#> 342 741 12.11163415
#> 343 742 12.10768293
#> 344 743 12.09139024
#> 345 744 12.06865854
#> 346 745 12.05143902
#> 347 746 12.05710000
#> 348 747 12.04326829
#> 349 748 12.01948780
#> 350 749 12.01843902
#> 351 750 11.97721951
#> 352 751 11.98568293
#> 353 752 11.99672500
#> 354 753 11.99153659
#> 355 754 11.98726829
#> 356 755 11.98925000
#> 357 756 11.98431707
#> 358 757 11.96502439
#> 359 758 11.94856098
#> 360 759 11.95110000
#> 361 760 11.93937500
#> 362 761 11.94153659
#> 363 762 11.93460976
#> 364 763 11.92241463
#> 365 764 11.91292683
#> 366 765 11.91960976
#> 367 766 11.90920000
#> 368 767 11.89360976
#> 369 768 11.90950000
#> 370 769 11.88200000
#> 371 770 11.90139024
#> 372 771 11.87229268
#> 373 772 11.88200000
#> 374 773 11.85577500
#> 375 774 11.86185000
#> 376 775 11.83456098
#> 377 776 11.82870732
#> 378 777 11.79068293
#> 379 778 11.73331707
#> 380 779 11.72260000
#> 381 780 11.74256098
#> 382 781 11.71546341
#> 383 782 11.68336585
#> 384 783 11.65256098
#> 385 784 11.62190000
#> 386 785 11.58375000
#> 387 786 11.56485000
#> 388 787 11.58065854
#> 389 788 11.51037500
#> 390 789 11.47070000
#> 391 790 11.43490244
#> 392 791 11.43137500
#> 393 792 11.42655000
#> 394 793 11.39025000
#> 395 794 11.37720000
#> 396 795 11.36345000
#> 397 796 11.31650000
#> 398 797 11.30500000
#> 399 798 11.27890244
#> 400 799 11.25535000
#> 401 800 11.22875000
#> 402 801 11.20130000
#> 403 802 11.18010000
#> 404 803 11.13355000
#> 405 804 11.09925000
#> 406 805 11.09660000
#> 407 806 11.07865000
#> 408 807 11.05790000
#> 409 808 11.01220000
#> 410 809 10.99755000
#> 411 810 10.96590000
#> 412 811 10.93051282
#> 413 812 10.93085000
#> 414 813 10.87740000
#> 415 814 10.86620000
#> 416 815 10.82860000
#> 417 816 10.78485000
#> 418 817 10.77164103
#> 419 818 10.72770000
#> 420 819 10.75950000
#> 421 820 10.75458974
#> 422 821 10.78300000
#> 423 822 10.77400000
#> 424 823 10.76825000
#> 425 824 10.76885000
#> 426 825 10.76935000
#> 427 826 10.73570000
#> 428 827 10.69222500
#> 429 828 10.62223077
#> 430 829 10.57050000
#> 431 830 10.51075000
#> 432 831 10.49605000
#> 433 832 10.46233333
#> 434 833 10.40587500
#> 435 834 10.39150000
#> 436 835 10.41022500
#> 437 836 10.38950000
#> 438 837 10.33980000
#> 439 838 10.32046154
#> 440 839 10.26615000
#> 441 840 10.17961538
#> 442 841 10.15650000
#> 443 842 10.12680000
#> 444 843 10.05720000
#> 445 844 10.02255000
#> 446 845 10.01656410
#> 447 846  9.96031250
#> 448 847  9.93246500
#> 449 848  9.92007692
#> 450 849  9.92555385
#> 451 850  9.89874615
#> 452 851  9.83922308
#> 453 852  9.78412308
#> 454 853  9.75693846
#> 455 854  9.72865385
#> 456 855  9.70964615
#> 457 856  9.72064615
#> 458 857  9.72046923
#> 459 858  9.69414615
#> 460 859  9.66870000
#> 461 860  9.67673500
#> 462 861  9.67470769
#> 463 862  9.62124359
#> 464 863  9.59161282
#> 465 864  9.53448718
#> 466 865  9.48221795
#> 467 866  9.42532308
#> 468 867  9.42196154
#> 469 868  9.41169231
#> 470 869  9.33900256
#> 471 870  9.34273500
#> 472 871  9.35832564
#> 473 872  9.32860000
#> 474 873  9.32593333
#> 475 874  9.30863846
#> 476 875  9.30343846
#> 477 876  9.31858462
#> 478 877  9.35144359
#> 479 878  9.37162308
#> 480 879  9.33505641
#> 481 880  9.39662308
#> 482 881  9.40768718
#> 483 882  9.43258205
#> 484 883  9.43595385
#> 485 884  9.45801538
#> 486 885  9.39966154
#> 487 886  9.34719231
#> 488 887  9.33137692
#> 489 888  9.20959231
#> 490 889  9.14444615
#> 491 890  9.07412308
#> 492 891  9.07272308
#> 493 892  8.99676154
#> 494 893  8.97930000
#> 495 894  8.99405897
#> 496 895  9.02053684
#> 497 896  9.03753590
#> 498 897  9.01471579
#> 499 898  9.03112821
#> 500 899  8.97823684
#> 501 900  8.95550769

getspec(system.file("testdata", package = "pavo"), ext = 'ttt')

#> 2 files found; importing spectra:

#>      wl avantes_export_long avantes_export
#> 1   300           13.624678         3.1487
#> 2   301            5.276307         3.1589
#> 3   302           11.023560         3.5700
#> 4   303           10.307208         3.9215
#> 5   304            8.980705         3.4034
#> 6   305            8.278200         3.7878
#> 7   306            9.208788         3.5876
#> 8   307           12.413864         4.0819
#> 9   308            7.878000         3.3291
#> 10  309            8.753368         3.4895
#> 11  310           11.325100         3.5958
#> 12  311           11.308051         3.2791
#> 13  312           11.806000         3.7085
#> 14  313           13.005017         3.1833
#> 15  314           13.445000         3.3174
#> 16  315           11.249267         3.1272
#> 17  316           11.846600         3.4802
#> 18  317           10.768883         3.2053
#> 19  318           11.304133         3.7435
#> 20  319           10.657750         3.5064
#> 21  320           10.128167         3.5004
#> 22  321            8.113640         3.7695
#> 23  322            7.936045         3.0782
#> 24  323            7.105346         3.6174
#> 25  324            7.995267         3.7919
#> 26  325           11.300900         3.8025
#> 27  326            8.664300         3.4200
#> 28  327            8.854781         3.6270
#> 29  328            9.942625         3.6022
#> 30  329           11.164167         3.3715
#> 31  330            9.333339         3.7531
#> 32  331           10.228905         3.5897
#> 33  332            8.529324         3.2939
#> 34  333            7.262425         3.4864
#> 35  334            8.113873         3.6349
#> 36  335            9.275302         3.0949
#> 37  336           10.992119         3.4867
#> 38  337           10.386067         3.5221
#> 39  338            9.720155         3.6964
#> 40  339            7.871376         3.4029
#> 41  340            8.426892         3.8356
#> 42  341            8.799900         3.9707
#> 43  342            8.122002         3.8189
#> 44  343            7.711958         3.8735
#> 45  344            9.007375         3.8944
#> 46  345            7.726317         3.9409
#> 47  346            7.432240         3.7156
#> 48  347            6.736702         3.5390
#> 49  348            7.721527         3.2439
#> 50  349            8.723975         3.5024
#> 51  350            8.626022         3.5816
#> 52  351           10.438700         4.0566
#> 53  352           10.158525         3.8713
#> 54  353            9.061749         3.9149
#> 55  354            8.295241         3.6999
#> 56  355            6.732814         3.5429
#> 57  356            6.407437         4.4038
#> 58  357            6.656164         3.7897
#> 59  358            8.540753         3.8220
#> 60  359            7.981805         4.0388
#> 61  360            8.200731         3.2907
#> 62  361            8.013527         3.6802
#> 63  362            8.018747         3.7239
#> 64  363            6.993866         3.6247
#> 65  364            6.948975         3.7148
#> 66  365            7.489102         3.8681
#> 67  366            7.989460         3.7169
#> 68  367            7.021139         3.6990
#> 69  368            7.589832         3.8144
#> 70  369            7.913195         3.8825
#> 71  370            7.447000         3.8171
#> 72  371            7.222630         3.7668
#> 73  372            7.176832         3.9765
#> 74  373            6.605175         3.7089
#> 75  374            6.285100         4.0894
#> 76  375            6.179647         3.6340
#> 77  376            6.614023         3.6828
#> 78  377            6.453842         4.0658
#> 79  378            6.711854         3.8002
#> 80  379            6.932268         3.9557
#> 81  380            6.968403         3.8227
#> 82  381            6.274686         3.8255
#> 83  382            6.190197         3.7804
#> 84  383            7.218693         3.9949
#> 85  384            6.932875         3.9315
#> 86  385            6.373603         3.9693
#> 87  386            6.005068         4.0217
#> 88  387            6.388133         3.8500
#> 89  388            6.663022         3.9628
#> 90  389            6.418460         4.0558
#> 91  390            6.388644         3.9115
#> 92  391            6.967334         4.0783
#> 93  392            6.819356         3.7879
#> 94  393            6.349461         3.8102
#> 95  394            6.256687         3.9478
#> 96  395            6.273366         4.4452
#> 97  396            6.553588         3.8187
#> 98  397            6.108805         4.0266
#> 99  398            5.898819         4.1396
#> 100 399            5.991895         4.0911
#> 101 400            5.639232         4.0899
#> 102 401            6.342615         3.9514
#> 103 402            6.147300         4.1269
#> 104 403            6.010692         4.2617
#> 105 404            6.333537         4.1683
#> 106 405            6.490569         4.0871
#> 107 406            6.241358         4.1305
#> 108 407            5.788260         4.1424
#> 109 408            6.157814         4.0957
#> 110 409            5.701488         4.2277
#> 111 410            5.940147         4.1021
#> 112 411            6.924328         4.1617
#> 113 412            6.087992         3.8605
#> 114 413            5.849449         4.3850
#> 115 414            5.683373         4.3590
#> 116 415            5.838020         4.3059
#> 117 416            5.885167         4.4339
#> 118 417            5.832951         4.2900
#> 119 418            6.246788         4.3030
#> 120 419            5.946205         4.2505
#> 121 420            5.753285         4.2643
#> 122 421            5.775020         4.3192
#> 123 422            6.060010         4.2917
#> 124 423            5.666710         4.4013
#> 125 424            5.300876         4.4192
#> 126 425            6.260776         4.4296
#> 127 426            5.927868         4.4799
#> 128 427            5.523244         4.3402
#> 129 428            5.680359         4.4549
#> 130 429            5.862654         4.3710
#> 131 430            6.005478         4.4217
#> 132 431            5.852803         4.3626
#> 133 432            5.633202         4.3898
#> 134 433            5.685012         4.4096
#> 135 434            5.652256         4.5058
#> 136 435            5.674714         4.5011
#> 137 436            5.632305         4.5701
#> 138 437            5.785119         4.6274
#> 139 438            5.828331         4.5987
#> 140 439            5.523895         4.5806
#> 141 440            5.687512         4.6629
#> 142 441            5.226607         4.6117
#> 143 442            5.539847         4.6966
#> 144 443            6.160193         4.6784
#> 145 444            5.510503         4.7690
#> 146 445            5.146653         4.8087
#> 147 446            5.039732         4.7808
#> 148 447            5.310500         4.8390
#> 149 448            5.649003         4.7986
#> 150 449            5.478061         4.8194
#> 151 450            5.218336         4.7875
#> 152 451            5.043793         4.7548
#> 153 452            5.435142         4.8210
#> 154 453            5.490605         4.9284
#> 155 454            5.834531         4.9248
#> 156 455            5.419097         5.0411
#> 157 456            6.047166         4.9363
#> 158 457            5.647617         5.0931
#> 159 458            5.406690         5.1138
#> 160 459            5.528110         5.0922
#> 161 460            5.847405         5.0276
#> 162 461            5.387800         5.1436
#> 163 462            5.063932         5.1816
#> 164 463            5.582700         5.1351
#> 165 464            5.462236         5.1034
#> 166 465            5.317108         5.1094
#> 167 466            5.637495         5.2988
#> 168 467            5.655116         5.0824
#> 169 468            5.531605         5.3410
#> 170 469            5.819490         5.3130
#> 171 470            5.806069         5.4010
#> 172 471            5.512176         5.2865
#> 173 472            5.273651         5.4104
#> 174 473            5.208234         5.5426
#> 175 474            5.351749         5.4377
#> 176 475            5.154792         5.5361
#> 177 476            5.388203         5.5596
#> 178 477            5.727675         5.5779
#> 179 478            5.950737         5.5648
#> 180 479            5.577909         5.5523
#> 181 480            5.477063         5.6653
#> 182 481            5.242231         5.6848
#> 183 482            5.374976         5.9212
#> 184 483            5.289807         5.8023
#> 185 484            5.091486         5.8743
#> 186 485            5.020537         5.9438
#> 187 486            5.142619         6.0388
#> 188 487            5.584985         5.9844
#> 189 488            5.175155         5.9544
#> 190 489            5.353825         5.9673
#> 191 490            5.709149         5.9115
#> 192 491            5.660500         6.0270
#> 193 492            5.618659         5.9752
#> 194 493            5.799031         6.2257
#> 195 494            5.698181         6.1121
#> 196 495            5.599755         6.2249
#> 197 496            5.196344         6.2125
#> 198 497            4.968624         6.2169
#> 199 498            5.025175         6.3501
#> 200 499            5.416758         6.2448
#> 201 500            5.636221         6.3877
#> 202 501            5.312793         6.3081
#> 203 502            5.644676         6.3649
#> 204 503            5.838385         6.4647
#> 205 504            5.874092         6.5877
#> 206 505            6.091984         6.5909
#> 207 506            5.812376         6.6324
#> 208 507            5.867271         6.4650
#> 209 508            5.420700         6.8351
#> 210 509            5.978405         6.8298
#> 211 510            5.888144         6.8113
#> 212 511            5.703202         6.9653
#> 213 512            5.575362         7.1235
#> 214 513            5.852566         7.0988
#> 215 514            5.670325         7.1244
#> 216 515            5.457093         7.1753
#> 217 516            5.783445         7.3012
#> 218 517            5.866984         7.3441
#> 219 518            5.780283         7.4322
#> 220 519            5.283485         7.2770
#> 221 520            5.763528         7.4944
#> 222 521            6.105460         7.5469
#> 223 522            5.756359         7.7945
#> 224 523            5.640051         7.7307
#> 225 524            5.741637         7.6106
#> 226 525            5.490468         7.7908
#> 227 526            5.645300         7.7892
#> 228 527            5.974745         7.8941
#> 229 528            6.130259         8.0406
#> 230 529            5.960550         8.0048
#> 231 530            5.735507         8.1355
#> 232 531            6.272810         8.2338
#> 233 532            5.971000         8.3604
#> 234 533            6.086697         8.4753
#> 235 534            6.097878         8.5502
#> 236 535            5.808900         8.5203
#> 237 536            6.324705         8.4856
#> 238 537            6.567258         8.4818
#> 239 538            6.111478         8.6828
#> 240 539            5.758576         8.6916
#> 241 540            5.573954         8.8087
#> 242 541            5.684269         9.1011
#> 243 542            5.885340         9.1397
#> 244 543            6.116086         9.1865
#> 245 544            6.157486         9.2986
#> 246 545            6.122153         9.3508
#> 247 546            6.166800         9.4986
#> 248 547            6.210924         9.7391
#> 249 548            6.387291         9.5321
#> 250 549            6.500802         9.6064
#> 251 550            6.146866         9.9329
#> 252 551            6.357066         9.7998
#> 253 552            6.270483         9.9119
#> 254 553            5.970298         9.8932
#> 255 554            6.027759        10.0706
#> 256 555            6.033157        10.2548
#> 257 556            6.327554        10.3564
#> 258 557            6.321679        10.4235
#> 259 558            6.668947        10.4461
#> 260 559            6.333391        10.6375
#> 261 560            6.617495        10.6926
#> 262 561            6.532407        10.8376
#> 263 562            6.489779        11.0381
#> 264 563            6.497647        11.0502
#> 265 564            6.442026        11.1467
#> 266 565            6.414534        11.2624
#> 267 566            6.404945        11.5495
#> 268 567            6.441373        11.5593
#> 269 568            6.443652        11.7231
#> 270 569            6.328114        11.7203
#> 271 570            6.278824        11.7975
#> 272 571            6.297617        11.8193
#> 273 572            6.390485        12.0314
#> 274 573            6.287605        12.1193
#> 275 574            6.205779        12.2115
#> 276 575            6.307461        12.2054
#> 277 576            6.305672        12.4790
#> 278 577            6.470824        12.5620
#> 279 578            6.539548        12.9507
#> 280 579            6.546455        12.7451
#> 281 580            6.481678        13.0597
#> 282 581            6.461022        13.0981
#> 283 582            6.601142        13.3829
#> 284 583            6.746986        13.5131
#> 285 584            6.800803        13.7167
#> 286 585            6.762490        13.6584
#> 287 586            6.805100        13.7166
#> 288 587            6.916595        13.6748
#> 289 588            6.887326        14.0659
#> 290 589            6.732036        14.3829
#> 291 590            6.811734        14.0865
#> 292 591            7.003436        14.4324
#> 293 592            6.978753        14.5572
#> 294 593            7.154686        14.5314
#> 295 594            7.123072        14.8969
#> 296 595            7.179503        14.7403
#> 297 596            7.316141        15.0934
#> 298 597            7.417503        15.1461
#> 299 598            7.401693        15.5567
#> 300 599            7.136207        15.6261
#> 301 600            7.343524        15.4423
#> 302 601            7.343493        16.0581
#> 303 602            7.251252        15.8785
#> 304 603            7.284148        16.0683
#> 305 604            7.480528        16.1915
#> 306 605            7.345131        16.1177
#> 307 606            7.423879        16.5828
#> 308 607            7.539031        16.4860
#> 309 608            7.581459        16.7972
#> 310 609            7.638531        16.8548
#> 311 610            7.588700        16.8859
#> 312 611            7.405976        16.9373
#> 313 612            7.378745        17.0873
#> 314 613            7.705817        16.9286
#> 315 614            7.885734        17.4187
#> 316 615            7.828431        17.3956
#> 317 616            7.852348        17.4448
#> 318 617            7.592586        17.8061
#> 319 618            7.748890        18.3132
#> 320 619            7.934916        18.1761
#> 321 620            7.915855        18.7688
#> 322 621            7.818072        18.4469
#> 323 622            8.104250        18.6282
#> 324 623            8.393653        18.6229
#> 325 624            8.217905        18.8278
#> 326 625            8.170200        19.2806
#> 327 626            8.427666        18.9704
#> 328 627            8.504376        19.7112
#> 329 628            8.316751        19.5547
#> 330 629            8.486852        19.6146
#> 331 630            8.355669        19.8695
#> 332 631            8.419453        19.9912
#> 333 632            8.862783        20.2501
#> 334 633            8.356841        20.5410
#> 335 634            8.619097        20.5744
#> 336 635            9.404854        20.7168
#> 337 636            8.901421        20.3630
#> 338 637            8.963674        21.2734
#> 339 638            8.817700        21.2380
#> 340 639            9.450852        20.9466
#> 341 640            9.289438        21.4761
#> 342 641            9.612581        21.4783
#> 343 642            9.880403        21.5209
#> 344 643            9.168682        22.0235
#> 345 644            9.469700        22.1964
#> 346 645            9.489038        23.1054
#> 347 646            9.644055        22.3284
#> 348 647           10.028072        23.0962
#> 349 648            9.656000        22.9967
#> 350 649            9.300962        23.5176
#> 351 650            9.571791        23.3030
#> 352 651            9.857688        23.2945
#> 353 652            9.648630        23.2722
#> 354 653            9.565310        23.6809
#> 355 654            9.218283        24.0806
#> 356 655            9.425198        24.1483
#> 357 656            9.491252        25.4356
#> 358 657            9.629784        24.4718
#> 359 658            9.836183        24.7032
#> 360 659            9.876944        24.4642
#> 361 660           10.255621        24.5189
#> 362 661            9.972626        24.9648
#> 363 662           10.129074        24.8506
#> 364 663           10.196828        25.2286
#> 365 664           10.259052        25.3255
#> 366 665           10.262158        25.0471
#> 367 666           10.283897        25.2522
#> 368 667           10.432690        25.7434
#> 369 668           10.332158        25.7718
#> 370 669           10.663842        25.9194
#> 371 670           10.796862        26.5442
#> 372 671           10.802632        27.0133
#> 373 672           10.693947        26.1726
#> 374 673           10.984053        26.5203
#> 375 674           11.174310        26.9380
#> 376 675           11.368397        27.1329
#> 377 676           10.995842        27.3850
#> 378 677           11.231000        27.2834
#> 379 678           11.364379        28.1359
#> 380 679           10.949448        27.7702
#> 381 680           10.774707        28.1869
#> 382 681           10.924158        27.7695
#> 383 682           11.423158        27.8909
#> 384 683           11.373982        28.3245
#> 385 684           11.475561        28.8958
#> 386 685           11.678086        28.9253
#> 387 686           11.400207        28.8044
#> 388 687           11.298684        29.1316
#> 389 688           11.363158        29.2402
#> 390 689           11.730586        29.2427
#> 391 690           11.847241        29.8739
#> 392 691           11.832810        29.4639
#> 393 692           11.578474        30.0884
#> 394 693           11.790259        30.4943
#> 395 694           11.940276        29.8330
#> 396 695           11.883207        30.1405
#> 397 696           11.932702        30.8067
#> 398 697           12.541672        30.8716
#> 399 698           12.686379        31.1003
#> 400 699           12.109579        30.8467
#> 401 700           12.489719        31.4912