Converts between irradiance and photon (quantum) flux

Some spectrometers will give illuminant values in units of irradiance (μWatt.cm-2), but physiological models require illuminants in units of photon (quantum) flux (μmol.s-1.m-2). The functions irrad2flux() and flux2irrad() allows for easy conversion of rspec objects between these units.

irrad2flux(rspecdata)

flux2irrad(rspecdata)

Arguments

rspecdata
(required) a rspec object containing illuminant values.

Value

a converted rspec object.

Author

Rafael Maia rm72@zips.uakron.edu