pavo: Perceptual Analysis, Visualization and Organization of Spectral Colour Data

A cohesive framework for parsing, analyzing and organizing colour from spectral data.

See also

Author

Maintainer: Thomas White thomas.white026@gmail.com (ORCID)

Authors:

Rafael Maia rafa.maia@gmail.com (ORCID)

Hugo Gruson hugo.gruson+R@normalesup.org (ORCID)

John Endler john.endler@deakin.edu.au

Chad Eliason cme16@zips.uakron.edu

Pierre-Paul Bitton bittonp@uwindsor.ca (ORCID)