Process images

Specify scales, resize, and/or define focal objects within images.

procimg(image, resize = NULL, rotate = NULL, scaledist = NULL,
  outline = FALSE, smooth = FALSE, iterations = 1L, col = "red",
  plotnew = FALSE, ...)

Arguments

image
(required) image data. Either a single image array, or a number of images stored in a list. Preferably the result of getimg.

resize
an integer specifying the scaling factor for linearly resizing images, if so desired. E.g. 0.5 to half the size of an image, or 2 to double it.

rotate
an integer specifying the angle of image rotation, in degrees. Images are rotated around the centre, and linearly interpolated.

scaledist
an integer, or numeric vector equal in length to the number of images, specifying the length of the scale in the image(s). Image(s) will then be presented, and the user asked to select either end of the scale corresponding to the input value.

outline
interactively specify the focal object in an image by clicking around its outline. The xy-coordinates of the resulting closed polygon are saved as an attribute, for use in generating a masking layer & separating animals/plants from backgrounds in further analyses. This is particularly useful when backgrounds are complex, such as in natural settings.

smooth
should the polygon specified when outline = TRUE be smoothed using Chaikin's corner-cuting algorithm? Defaults to FALSE.

iterations
the number of smoothing iterations, when smooth = TRUE. Defaults to 1.

col
the color of the marker points and/or line, when using interactive options.

plotnew
should plots be opened in a new window? Defaults to FALSE.

...
additional graphical parameters. Also see par.

Value

an image, or list of images, for use in further pavo functions.

References

Chaikin, G. 1974. An algorithm for high speed curve generation. Computer Graphics and Image Processing 3, 346-349.

Examples

# NOT RUN {
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio <- procimg(papilio, scaledist = 10)

# Assign individual scales to each image, after slightly reducing their size.
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes <- procimg(snakes, scaledist = c(10, 14), resize = 0.95)
# }