Hue projection plot

Produces a 2D projection plot of points in a color space

Adds points to a tetrahedral colorspace projection

projplot(tcsdata, ...)

projpoints(tcsres, ...)

Arguments

tcsdata
(required) color space coordinates, possibly a result from the tcs function, containing values for the 'h.theta' and 'h.phi' coordinates as columns (labeled as such).

...
additonal parameters to be passed to the plotting of data points.

tcsres
(required) color space coordinates, possibly a result from the tcs function, containing values for the 'h.theta' and 'h.phi' coordinates as columns (labeled as such).

Value

projplot creates a 2D plot of color points projected from the tetrahedron to its encapsulating sphere, and is ideal to visualize differences in hue. projpoints creates points in a projection color space plot produced by projplot.

Note

projplot uses the Mollweide projection, and not the Robinson projection, which has been used in the past. Among other advantages, the Mollweide projection preserves area relationships within latitudes without distortion.

References

Stoddard, M. C., & Prum, R. O. (2008). Evolution of avian plumage color in a tetrahedral color space: A phylogenetic analysis of new world buntings. The American Naturalist, 171(6), 755-776.

Endler, J. A., Westcott, D., Madden, J., & Robson, T. (2005). Animal visual systems and the evolution of color patterns: Sensory processing illuminates signal evolution. Evolution, 59(8), 1795-1818.

Examples

not_run({
  data(sicalis)
  vis.sicalis <- vismodel(sicalis, visual = 'avg.uv')
  tcs.sicalis <- colspace(vis.sicalis, space = 'tcs')
  projplot(tcs.sicalis, pch = 16, col = setNames(rep(1:3, 7), rep(c('C', 'T', 'B'), 7)))
})