Angle-resolved reflectance data for the iridescent wing patch of a male green-winged teal (Anas carolinensis)

Dataset containing reflectance measurements from the wing patch of a single male at different incident angles (15-75 degrees in 5-degree increments).

Usage

data(teal)

Format

An object of class rspec (inherits from data.frame) with 401 rows and 13 columns.

Author

Chad Eliason cme16@zips.uakron.edu