Bootstrap colour distance confidence intervals

Uses a bootstrap procedure to generate confidence intervals for the mean colour distance between two or more samples of colours

bootcoldist(vismodeldata, by, boot.n = 1000, alpha = 0.95, cores = NULL, ...)

Arguments

vismodeldata
(required) quantum catch colour data. Can be the result from vismodel(), or colspace(). Data may also be independently calculated quantum catches, in the form of a data frame with columns representing photoreceptors.

by
(required) a numeric or character vector indicating the group to which each row from the object belongs to.

boot.n
number of bootstrap replicates (defaults to 1000)

alpha
the confidence level for the confidence intervals (defaults to 0.95)

cores
deprecated. See future::plan() for more details on how to customise your parallelisation strategy.

...
other arguments to be passed to coldist(). Must at minimum include n and weber. See coldist() for details.

Value

a matrix including the empirical mean and bootstrapped confidence limits for dS (and dL if achromatic = TRUE).

Details

You can customise the type of parallel processing used by this function with the future::plan() function. This works on all operating systems, as well as high performance computing (HPC) environment. Similarly, you can customise the way progress is shown with the progressr::handlers() functions (progress bar, acoustic feedback, nothing, etc.)

References

Maia, R., White, T. E., (2018) Comparing colors using visual models. Behavioral Ecology, ary017 doi: 10.1093/beheco/ary017

Examples

if (FALSE) {
data(sicalis)
vm <- vismodel(sicalis, achromatic = "bt.dc")
gr <- gsub("ind..", "", rownames(vm))
bootcoldist(vm, by = gr, n = c(1, 2, 2, 4), weber = 0.1, weber.achro = 0.1)
}