CIE colour spaces

Calculates coordinates and colorimetric variables that represent reflectance spectra in either the CIEXYZ (1931), CIELAB (1971), or CIELCH (1971) colourspaces.

cie(vismodeldata, space = c("XYZ", "LAB", "LCh"))

Arguments

vismodeldata
(required) quantum catch color data. Can be either the result from vismodel or independently calculated data (in the form of a data frame with three columns representing trichromatic viewer).

space
(required) Choice between XYZ (1931), LAB (1971), or LCH colour models.

Value

Object of class colspace containing:

X, Y, Z: Tristimulus values.

x, y, z: Cartesian coordinates, when using space = XYZ.

L, a, b: Lightness, L, and colour-opponent a (redness-greenness) and b (yellowness-blueness) values, in a Cartesian coordinate space. Returned when using space = LAB.

L, C, h: Lightness, L, chroma C and hue-angle h (degrees), the latter of which are cylindrical representations of a and b from the CIELAB model. Returned when using space = LCh.

References

Smith T, Guild J. (1932) The CIE colorimetric standards and their use. Transactions of the Optical Society, 33(3), 73-134.

Westland S, Ripamonti C, Cheung V. (2012). Computational colour science using MATLAB. John Wiley & Sons.

Stockman, A., & Sharpe, L. T. (2000). Spectral sensitivities of the middle- and long-wavelength sensitive cones derived from measurements in observers of known genotype. Vision Research, 40, 1711-1737.

CIE (2006). Fundamental chromaticity diagram with physiological axes. Parts 1 and 2. Technical Report 170-1. Vienna: Central Bureau of the Commission Internationale de l Eclairage.

Examples

not_run({
  data(flowers)
  vis.flowers <- vismodel(flowers, visual = 'cie10', illum = 'D65')
  flowers.ciexyz <- colspace(vis.flowers, space = 'ciexyz')
  flowers.cielab <- colspace(vis.flowers, space = 'cielab')
  flowers.cielch <- colspace(vis.flowers, space = 'cielch')
})