Identify colour classes in an image for adjacency analyses

Use k-means clustering to classify image pixels into discrete colour classes.

classify(imgdat, kcols = NULL, refID = NULL, interactive = FALSE,
  plotnew = FALSE, col = "red", cores = getOption("mc.cores", 2L),
  ...)

Arguments

imgdat
(required) image data. Either a single image, or a series of images stored in a list. Preferably the result of getimg.

kcols
the number of discrete colour classes present in the input image(s). Can be a single integer when only a single image is present, or if kcols is identical for all images. When passing a list of images, kcols can also be a vector the same length as imgdat, or a data.frame with two columns specifying image file names and corresponding kcols. This argument can optionally be disregarded when interactive = TRUE, and kcols will be inferred from the number of selections.

refID
the optional numeric index of a 'reference' image, for use when passing a list of images. Other images will be k-means classified using centres identified in the single reference image, thus helping to ensure that homologous pattern elements will be reliably classified between images, if so desired.

interactive
interactively specify the colour-category 'centers', for k-means clustering. When TRUE, the user is asked to click a number of points (equal to kcols, if specified, otherwise user-determined) that represent the distinct colours of interest. If a reference image is specified, it will be the only image presented.

plotnew
Should plots be opened in a new window when interactive = TRUE? Defaults to FALSE.

col
the color of the marker points, when interactive = TRUE.

cores
number of cores to be used in parallel processing. If 1, parallel computing will not be used. Defaults to getOption("mc.cores", 2L). Not available on Windows.

...
additional graphical parameters when interactive = TRUE. Also see par.

Value

A matrix, or list of matrices, of class rimg containing the colour class classifications ID at each pixel location. The RGB values corresponding to k-means centres (i.e. colour classes) are stored as object attributes.

Note

Since the kmeans process draws on random numbers to find initial cluster centres when interactive = FALSE, use set.seed if reproducible cluster ID's are desired between runs.

Examples

# NOT RUN {
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))
papilio_class <- classify(papilio, kcols = 4)

# Multiple images, with interactive classification and a reference image
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
snakes_class <- classify(snakes, refID = 1, interactive = TRUE)
# }