Import image data

Finds and imports PNG, JPEG, and/or BMP images.

getimg(imgpath = getwd(), subdir = FALSE, max.size = 1,
  cores = getOption("mc.cores", 2L))

Arguments

imgpath
(required) either the full path to a given image (including extension), or the path to a folder in which multiple image files are located. Mixed file formats within a folder are fine.

subdir
should subdirectories within the where folder be included in the search? (defaults to FALSE).

max.size
maximum size of all images to be allowed in memory, in GB. Defaults to 2.

cores
number of cores to be used in parallel processing. If 1, or if totale image sizes exceed 200 mb in memory, parallel computing will not be used. Defaults to getOption("mc.cores", 2L).

Value

a array, or list of arrays, of class rimg, containing image data for use in further pavo functions.

Examples

# NOT RUN {
# Single image
papilio <- getimg(system.file("testdata/images/papilio.png", package = 'pavo'))

# Multiple images
snakes <- getimg(system.file("testdata/images/snakes", package = 'pavo'))
# }