Import spectra files

Finds and imports spectra files from a folder. Currently works for reflectance files generated in Ocean Optics SpectraSuite (USB2000, USB4000 and Jaz spectrometers), CRAIC software (after exporting) and Avantes (before or after exporting).

getspec(where = getwd(), ext = "txt", lim = c(300, 700),
  decimal = ".", sep = NULL, subdir = FALSE, subdir.names = FALSE,
  cores = getOption("mc.cores", 2L), ignore.case = TRUE)

Arguments

where
(required) folder in which files are located.

ext
file extension to be searched for, without the "." (defaults to "txt").

lim
a vector with two numbers determining the wavelength limits to be considered (defaults to 300 and 700).

decimal
character to be used to identify decimal plates (defaults to ".").

sep
column delimiting characters to be considered in addition to the default (which are: tab, space, and ";")

subdir
should subdirectories within the where folder be included in the search? (defaults to FALSE).

subdir.names
should subdirectory path be included in the name of the spectra? (defaults to FALSE).

cores
Number of cores to be used. If greater than 1, import will use parallel processing (not available in Windows).

ignore.case
Logical. Should the extension search be case insensitive? (defaults to `TRUE``)

Value

A data frame, of class rspec, containing individual imported spectral files as columns. Reflectance values are interpolated to the nearest wavelength integer.

References

Montgomerie R (2006) Analyzing colors. In: Hill G, McGraw K (eds) Bird coloration. Harvard University Press, Cambridge, pp 90-147.

Examples

rspecdata <- getspec(system.file("testdata", package = "pavo"), ext = "ttt", lim = c(400, 700))

#>    Length     Class      Mode 
#>         2 character character 

#> 2 files found; importing spectra:

head(rspecdata)

#>    wl avantes_export avantes_export_long
#> 1 400         4.0899            5.639232
#> 2 401         3.9514            6.342615
#> 3 402         4.1269            6.147300
#> 4 403         4.2617            6.010692
#> 5 404         4.1683            6.333537
#> 6 405         4.0871            6.490569