Rotate Cartesian coordinates obtained from jnd2xyz()

Rotate Cartesian coordinates obtained from jnd2xyz()

jndrot(jnd2xyzres, center = c("mean", "achro"), ref1 = "l",
  ref2 = "u", axis1 = c(1, 1, 0), axis2 = c(0, 0, 1))

Arguments

jnd2xyzres
(required) the output from a jnd2xyz() call.

center
should the vectors for rotation be centered in the achromatic center ("achro") or the data centroid ("mean", the default)?

ref1
the cone to be used as a the first reference. May be NULL (for no first rotation in the 3-dimensional case) or must match name in the original data that was used for coldist(). Defaults to 'l'.

ref2
the cone to be used as a the second reference.May be NULL (for no first rotation in the 3-dimensional case) or must match name in the original data that was used for coldist(). Defaults to 'u'. (only used if data has 3 dimensions).

axis1
A vector of length 3 composed of 0's and 1's, with 1's representing the axes (x,y,z) to rotate around. Defaults to c(1,1,0), such that the rotation aligns with the xy plane (only used if data has 2 or 3 dimensions). Ignored if ref1 is `NULL`` (in 3-dimensional case only)

axis2
A vector of length 3 composed of 0's and 1's, with 1's representing the axes (x,y,z) to rotate around. Defaults to c(0,0,1), such that the rotation aligns with the z axis (only used if data has 3 dimensions). Ignored if ref2 is `NULL`` (in 3-dimensional case only)

Examples

data(flowers)
vis.flowers <- vismodel(flowers)
cd.flowers <- coldist(vis.flowers)

#> Warning: Quantum catch are relative, distances may not be meaningful

jndrot(jnd2xyz(cd.flowers))

#>                                    x           y           z
#> Goodenia_heterophylla      0.5424525 -5.64780370   7.0514924
#> Goodenia_geniculata       -1.0550989 -2.50605704  -6.4010132
#> Goodenia_gracilis         -3.2108736 12.34251010 -22.9160486
#> Xyris_operculata           2.8777939 -1.88599435  -3.8107457
#> Eucalyptus_sp             -0.6213975 -3.26571043   3.2303526
#> Faradaya_splendida        -1.1936353  0.68528547  -2.7868330
#> Gaultheria_hispida         1.7259887 -4.53631733  -1.9683611
#> Geitonoplesium_cymosum     7.2165208 -8.46082209   8.3554048
#> Euryomyrtus_ramosissima   -2.0871030 -0.00175431   5.1435291
#> Genista_linifolia         -0.8989973 -1.67855884  -2.3301504
#> Genista_monspessulana     -0.7098904 -3.12423952  -2.8881030
#> Geranium_sp                8.8675762  5.29282374  11.3644393
#> Glycine_clandestina       -0.6209529 -2.75483782  -4.9021716
#> Gompholobium_ecostatum_1   2.2856298 -0.39896279   5.9004073
#> Gompholobium_ecostatum_2  -0.7407280 -2.09938884  -6.2491817
#> Gompholobium_grandiflorum -2.3373318  5.12665208  -6.6125576
#> Gompholobium_huegelii      0.4043442 -1.96264953   6.5006043
#> Gompholobium_virgatum      1.3965806 -1.53710744   6.2822410
#> Gonocarpus_humilis        -0.5781891 -3.88455026  -3.4985432
#> Gonocarpus_teucrioides    -0.5000199 -2.16889995   2.3336963
#> Hibbertia_obtusifolia     -0.7777395 -1.33732709  -1.8375439
#> Zieria_arborescens        -2.9397129 10.24687737   9.4440349
#> Goodenia_lanata           -3.0651433  8.49316930 -13.6100744
#> Goodenia_ovata            -1.5925659 11.29769118  10.3387120
#> Goodenia_rotundifolia      1.7894663 -6.87365069   6.1896101
#> Grevillea_buxifolia        1.9016722 -4.04662703   1.6394104
#> Grevillea_steiglitziana   -3.6841457  6.10965035  -0.6206625
#> Grevillea_oleoides        -0.6650504 -3.11383460  -4.3953309
#> Gymnostachys_anceps       -2.9254163 11.26978172  -9.2576390
#> Hakea_actites              0.1172613 -6.58872224   8.7478471
#> Hardenbergia_violaceae    -1.4387104  0.81645128  -3.1446578
#> Hibbertia_acicularis      -0.5814491 -3.50513460   2.0046822
#> Hibbertia_bracteata        7.0249422 -3.76843830   1.5050828
#> Hibbertia_empetrifolia    -2.4727661  8.99013129   9.5245585
#> Hibbertia_procumbens      -0.5037057 -3.44357995  -0.6629848
#> Hibbertia_linearis        -0.9496057 -2.08005517  -7.6635029