Subset rspec, vismodel, and colspace objects

Subsets various object types based on a given vector or grep partial matching of data names.

# S3 method for rspec
subset(x, subset, ...)

# S3 method for colspace
subset(x, subset, ...)

# S3 method for vismodel
subset(x, subset, ...)

Arguments

x
(required) an object of class rspec, vismodel, or colspace, containing spectra, visual model output or colorspace data to subset.

subset
a string used for partial matching of observations.

...
class consistency (ignored).

Value

a subsetted object of the same class as the input object.

Examples

not_run({
  data(sicalis)
  vis.sicalis <- vismodel(sicalis)
  tcs.sicalis <- colspace(vis.sicalis, space = 'tcs')

  # Subset all 'crown' patches (C in file names)
  head(subset(sicalis, "C"))
  subset(vis.sicalis, "C")
  subset(tcs.sicalis, "C")[, 1:5]
})