Colorspace data summary

Returns the attributes of colspace objects.

# S3 method for colspace
summary(object, by = NULL, ...)

Arguments

object
(required) a colspace object.

by
when the input is in tcs colorspace, by is either a single value specifying the range of color points for which summary tetrahedral-colorspace variables should be calculated (for example, by = 3 indicates summary will be calculated for groups of 3 consecutive color points (rows) in the quantum catch color data frame) or a vector containing identifications for the rows in the quantum catch color data frame (in which case summaries will be calculated for each group of points sharing the same identification). If by is left blank, the summary statistics are calculated across all color points in the data.

...
class consistency (ignored).

Value

returns all attributes of the data as mapped to the selected colourspace, including options specified when calculating the visual model. Also return the default data.frame summary, except when the object is the result of tcs, in which case the following variables are output instead:

centroid.u, .s, .m, .l the centroids of usml coordinates of points.

c.vol the total volume occupied by the points.

rel.c.vol volume occupied by the points relative to the tetrahedron volume.

colspan.m the mean hue span.

colspan.v the variance in hue span.

huedisp.m the mean hue disparity.

huedisp.v the variance in hue disparity.

mean.ra mean saturation.

max.ra maximum saturation achieved by the group of points.

References

Stoddard, M. C., & Prum, R. O. (2008). Evolution of avian plumage color in a tetrahedral color space: A phylogenetic analysis of new world buntings. The American Naturalist, 171(6), 755-776.

Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.

Examples

# NOT RUN {
# Colour hexagon
data(flowers)
vis.flowers <- vismodel(flowers, visual = 'apis', qcatch = 'Ei', relative = FALSE,
                        vonkries = TRUE, bkg = 'green')
flowers.hex <- hexagon(vis.flowers)
summary(flowers.hex)

# Tetrahedral model
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual='avg.uv')
csp.sicalis <- colspace(vis.sicalis)
summary(csp.sicalis, by = rep(c('C', 'T', 'B'), 7))
# }