Colourspace data summary

Returns the attributes of colspace objects.

# S3 method for colspace
summary(object, by = NULL, ...)

Arguments

object
(required) a colspace object.

by
when the input is in tcs colourspace, by is either a single value specifying the range of colour points for which summary tetrahedral-colourspace variables should be calculated (for example, by = 3 indicates summary will be calculated for groups of 3 consecutive colour points (rows) in the quantum catch colour data frame) or a vector containing identifications for the rows in the quantum catch colour data frame (in which case summaries will be calculated for each group of points sharing the same identification). If by is left blank, the summary statistics are calculated across all colour points in the data.

...
class consistency (ignored).

Value

returns all attributes of the data as mapped to the selected colourspace, including options specified when calculating the visual model. Also return the default data.frame summary, except when the object is the result of tcspace(), in which case the following variables are output instead:

centroid.u, .s, .m, .l the centroids of usml coordinates of points.

c.vol the total volume occupied by the points.

rel.c.vol volume occupied by the points relative to the tetrahedron volume.

colspan.m the mean hue span.

colspan.v the variance in hue span.

huedisp.m the mean hue disparity.

huedisp.v the variance in hue disparity.

mean.ra mean saturation.

max.ra maximum saturation achieved by the group of points.

References

Stoddard, M. C., & Prum, R. O. (2008). Evolution of avian plumage color in a tetrahedral color space: A phylogenetic analysis of new world buntings. The American Naturalist, 171(6), 755-776.

Endler, J. A., & Mielke, P. (2005). Comparing entire colour patterns as birds see them. Biological Journal Of The Linnean Society, 86(4), 405-431.

Examples

# Colour hexagon
data(flowers)
vis.flowers <- vismodel(flowers,
  visual = "apis", qcatch = "Ei", relative = FALSE,
  vonkries = TRUE, bkg = "green"
)
flowers.hex <- hexagon(vis.flowers)
summary(flowers.hex)

#> Colorspace & visual model options:
#>  * Colorspace: hexagon 
#>  * Quantal catch: Ei 
#>  * Visual system, chromatic: apis 
#>  * Visual system, achromatic: none 
#>  * Illuminant: ideal, scale = 1 (von Kries colour correction applied) 
#>  * Background: green 
#>  * Relative: FALSE 
#>  

# Tetrahedral model
data(sicalis)
vis.sicalis <- vismodel(sicalis, visual = "avg.uv")
csp.sicalis <- colspace(vis.sicalis)
summary(csp.sicalis, by = rep(c("C", "T", "B"), 7))

#> Colorspace & visual model options:
#>  * Colorspace: tcs 
#>  * Quantal catch: Qi 
#>  * Visual system, chromatic: avg.uv 
#>  * Visual system, achromatic: none 
#>  * Illuminant: ideal, scale = 1 (von Kries colour correction not applied) 
#>  * Background: ideal 
#>  * Relative: TRUE 
#>  

#>   centroid.u centroid.s centroid.m centroid.l        c.vol    rel.c.vol
#> B 0.14091298 0.04946432  0.3838526  0.4257701 6.281511e-06 2.901306e-05
#> C 0.06947461 0.03144895  0.4054651  0.4936114 4.739152e-06 2.188920e-05
#> T 0.15368451 0.06413428  0.3766734  0.4055078 5.183721e-06 2.394258e-05
#>    colspan.m    colspan.v  huedisp.m    huedisp.v   mean.ra    max.ra
#> B 0.05758429 0.0013841927 0.06717740 0.0011466898 0.8021427 0.9039261
#> C 0.03193253 0.0003263454 0.06164553 0.0013887690 0.8742042 0.9061528
#> T 0.06171418 0.0012215063 0.05595025 0.0005378623 0.7434629 0.8816377